Extract OperationMetadata
