+++ src/components/openapi/OperationMetadata.module.css
@@ -1,0 +1,12 @@
+.operationIdValue > div {
+  display: flex;
+  align-items: center;
+}
+
+.operationIdValue > div > div {
+  flex-grow: 1;
+}
+
+.labelWithHelp {
+  display: flex;
+}
